compiltation issue

+++ spotiflow/lib/spotflow2d.cpp
@@ -124,7 +124,7 @@ static PyObject *c_gaussian2d(PyObject *self, PyObject *args)
 
             // get the closest point
             const float query_pt[2] = {(float)j, (float)i};
-            unsigned long ret_index;
+            size_t ret_index;
             float out_dist_sqr;
 
             index.knnSearch(
